+++ src/view/_loader.py
@@ -153,7 +153,8 @@ def _format_body(
         vbody_defaults[k] = default
 
     return [
-        (TYPECODE_CLASSTYPES, k, v, vbody_defaults[k]) for k, v in vbody_final.items()
+        (TYPECODE_CLASSTYPES, k, v, vbody_defaults[k])
+        for k, v in vbody_final.items()
     ]
 
 
@@ -343,7 +344,9 @@ def _build_type_codes(
                 vbody_types = vbody
 
             doc = {}
-            codes.append((TYPECODE_CLASS, tp, _format_body(vbody_types, doc, tp)))
+            codes.append(
+                (TYPECODE_CLASS, tp, _format_body(vbody_types, doc, tp))
+            )
             setattr(tp, "_view_doc", doc)
             continue
 
@@ -357,7 +360,9 @@ def _build_type_codes(
             key, value = get_args(tp)
 
             if key is not str:
-                raise InvalidBodyError(f"dictionary keys must be strings, not {key}")
+                raise InvalidBodyError(
+                    f"dictionary keys must be strings, not {key}"
+                )
 
             tp_codes = _build_type_codes((value,))
             codes.append((TYPECODE_DICT, None, tp_codes))
@@ -418,6 +423,10 @@ def finalize(routes: list[Route], app: ViewApp):
 
     for route in routes:
         set_load(route)
+        route.app = app
+
+        if route.parallel_build is None:
+            route.parallel_build = app.config.build.parallel
 
         if route.method:
             target = targets[route.method]
@@ -456,7 +465,9 @@ def finalize(routes: list[Route], app: ViewApp):
                     route.inputs.insert(index, 1)
                     continue
 
-                default = v.default if v.default is not inspect._empty else _NoDefault
+                default = (
+                    v.default if v.default is not inspect._empty else _NoDefault
+                )
 
                 route.inputs.insert(
                     index,
@@ -483,12 +494,11 @@ def finalize(routes: list[Route], app: ViewApp):
         if target:
             target(
                 route.path,  # type: ignore
-                route.func,
+                route,
                 route.cache_rate,
                 _format_inputs(route.inputs),
                 route.errors or {},
                 route.parts,  # type: ignore
-                [i for i in reversed(route.middleware_funcs)],
             )
         else:
             for i in (route.method_list) or targets.keys():
@@ -500,7 +510,6 @@ def finalize(routes: list[Route], app: ViewApp):
                     _format_inputs(route.inputs),
                     route.errors or {},
                     route.parts,  # type: ignore
-                    [i for i in reversed(route.middleware_funcs)],
                 )
 
 
@@ -552,7 +561,9 @@ def load_fs(app: ViewApp, target_dir: Path) -> None:
                     )
                 else:
                     path_obj = Path(path)
-                    stripped = list(path_obj.parts[len(target_dir.parts) :])  # noqa
+                    stripped = list(
+                        path_obj.parts[len(target_dir.parts) :]
+                    )  # noqa
                     if stripped[-1] == "index.py":
                         stripped.pop(len(stripped) - 1)
 
@@ -605,7 +616,8 @@ def load_simple(app: ViewApp, target_dir: Path) -> None:
             for route in mini_routes:
                 if not route.path:
                     raise InvalidRouteError(
-                        "omitting path is only supported" " on filesystem loading",
+                        "omitting path is only supported"
+                        " on filesystem loading",
                     )
 
                 routes.append(route)