+++ src/view/routing.py
@@ -1,18 +1,26 @@
 from __future__ import annotations
 
+import asyncio
 import builtins
 import inspect
 import re
+from collections.abc import Coroutine
 from contextlib import suppress
 from dataclasses import dataclass, field
 from enum import Enum
 from typing import (Any, Callable, Generic, Iterable, Literal, Type, TypeVar,
                     Union)
 
+from typing_extensions import ParamSpec, TypeAlias
+
 from ._util import LoadChecker, make_hint
+from .build import run_step
 from .exceptions import InvalidRouteError, MistakeError
-from .typing import (Middleware, StrMethod, Validator, ValueType, ViewResponse,
-                     ViewRoute)
+from .typing import (TYPE_CHECKING, Middleware, StrMethod, Validator,
+                     ValueType, ViewResult, ViewRoute)
+
+if TYPE_CHECKING:
+    from .app import App
 
 __all__ = (
     "get",
@@ -44,6 +52,7 @@ class Method(Enum):
 
 
 V = TypeVar("V", bound="ValueType")
+P = ParamSpec("P")
 
 
 @dataclass
@@ -72,10 +81,10 @@ RouteData = Literal[1, 2]
 
 
 @dataclass
-class Route(LoadChecker):
+class Route(Generic[P], LoadChecker):
     """Standard Route Wrapper"""
 
-    func: Callable[..., ViewResponse]
+    func: ViewRoute[P]
     path: str | None
     method: Method | None
     inputs: list[RouteInput | RouteData]
@@ -84,8 +93,13 @@ class Route(LoadChecker):
     errors: dict[int, ViewRoute] | None = None
     extra_types: dict[str, Any] = field(default_factory=dict)
     parts: list[str | Part[Any]] = field(default_factory=list)
-    middleware_funcs: list[Middleware] = field(default_factory=list)
+    middleware_funcs: list[Middleware[P]] = field(default_factory=list)
     method_list: list[Method] | None = None
+    steps: Iterable[str] | None = None
+    parallel_build: bool | None = None
+    app: App | None = (
+        None  # This will only be non-None after the loader has been called.
+    )
 
     def error(self, status_code: int):
         def wrapper(handler: ViewRoute):
@@ -97,10 +111,10 @@ class Route(LoadChecker):
 
         return wrapper
 
-    def middleware(self, func_or_none: Middleware | None = None):
+    def middleware(self, func_or_none: Middleware[P] | None = None):
         """Define a middleware function for the route."""
 
-        def inner(func: Middleware):
+        def inner(func: Middleware[P]) -> None:
             self.middleware_funcs.append(func)
 
         if func_or_none:
@@ -113,14 +127,49 @@ class Route(LoadChecker):
 
     __str__ = __repr__
 
-    def __call__(self, *args: Any, **kwargs: Any) -> Any:
-        return self.func(*args, **kwargs)
+    async def __call__(self, *args: P.args, **kwargs: P.kwargs) -> ViewResult:
+        coros: list[Coroutine] = []
+        assert self.app, "app is None"
+
+        for step in self.steps or ():
+            coros.append(run_step(self.app, step))
+
+        if self.parallel_build:
+            await asyncio.gather(*coros)
+
+        if self.middleware_funcs:
+            index = len(self.middleware_funcs) - 1
+            mw = self.middleware_funcs[index]
+
+            async def call_next() -> ViewResult:
+                nonlocal index
+                index -= 1
+
+                if index < 0:
+                    func = self.func(*args, **kwargs)
+
+                    if isinstance(func, Coroutine):
+                        return await func
+
+                    return func  # type: ignore
+
+                mw = self.middleware_funcs[index]
+                return await mw(call_next, *args, **kwargs)
+
+            return await mw(call_next, *args, **kwargs)
+
+        result = self.func(*args, **kwargs)
+        if isinstance(result, Coroutine):
+            return await result
+
+        # type checker still thinks its async for some reason
+        return result  # type: ignore
 
 
-RouteOrCallable = Union[Route, ViewRoute]
+RouteOrCallable: TypeAlias = Union[Route[P], ViewRoute[P]]
 
 
-def _ensure_route(r: RouteOrCallable) -> Route:
+def _ensure_route(r: RouteOrCallable[P]) -> Route[P]:
     if isinstance(r, Route):
         return r
 
@@ -128,7 +177,7 @@ def _ensure_route(r: RouteOrCallable) -> Route:
 
 
 def route_types(
-    r: RouteOrCallable,
+    r: RouteOrCallable[P],
     data: type[Any] | tuple[type[Any]] | dict[str, Any],
 ) -> Route:
     route = _ensure_route(r)
@@ -141,14 +190,18 @@ def route_types(
         route.extra_types[data.__name__] = data
     else:
         raise InvalidRouteError(
-            "expected type, tuple of tuples," f" or a dict, got {type(data).__name__}"
+            "expected type, tuple of tuples,"
+            f" or a dict, got {type(data).__name__}"
         )
 
     return route
 
 
+_DefinedByConfig = None
+
+
 def _method(
-    r: RouteOrCallable,
+    r: RouteOrCallable[P],
     raw_path: str | None,
     doc: str | None,
     method: Method | None,
@@ -156,23 +209,31 @@ def _method(
     *,
     method_list: list[Method] | None = None,
     inputs: list[RouteInput | RouteData] | None = None,
-) -> Route:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Route[P]:
     route = _ensure_route(r)
     route.method = method
     route.cache_rate = cache_rate
     route.method_list = method_list
+    route.steps = steps
+    route.parallel_build = parallel_build
     util_path = raw_path or "/"
 
     if not util_path.startswith("/"):
         raise MistakeError(
             "paths must started with a slash",
-            hint=make_hint(f'This should be "/{util_path}" instead', back_lines=2),
+            hint=make_hint(
+                f'This should be "/{util_path}" instead', back_lines=2
+            ),
         )
 
     if util_path.endswith("/") and (len(util_path) != 1):
         raise MistakeError(
             "paths must not end with a slash",
-            hint=make_hint(f'This should be "{util_path[:-1]}" instead', back_lines=2),
+            hint=make_hint(
+                f'This should be "{util_path[:-1]}" instead', back_lines=2
+            ),
         )
 
     if "{" in util_path:
@@ -222,18 +283,20 @@ def _method(
     return route
 
 
-Path = Callable[[RouteOrCallable], Route]
+Path: TypeAlias = Callable[[RouteOrCallable[P]], Route[P]]
 
 
 def _method_wrapper(
-    path_or_route: str | None | RouteOrCallable,
+    path_or_route: str | None | RouteOrCallable[P],
     doc: str | None,
     method: Method | None,
     cache_rate: int,
     *,
     method_list: list[Method] | None = None,
     inputs: list[RouteInput | RouteData] | None = None,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     def inner(r: RouteOrCallable) -> Route:
         if (not isinstance(path_or_route, str)) and path_or_route:
             raise TypeError(f"{path_or_route!r} is not a string")
@@ -246,6 +309,8 @@ def _method_wrapper(
             cache_rate,
             method_list=method_list,
             inputs=inputs,
+            steps=steps,
+            parallel_build=parallel_build,
         )
 
     if not path_or_route:
@@ -258,17 +323,21 @@ def _method_wrapper(
 
 
 def get(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a GET route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -279,21 +348,32 @@ def get(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.GET, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.GET,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def post(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a POST route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -304,21 +384,32 @@ def post(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.POST, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.POST,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def patch(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a PATCH route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -329,21 +420,32 @@ def patch(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.PATCH, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.PATCH,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def put(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a PUT route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -354,21 +456,32 @@ def put(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.PUT, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.PUT,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def delete(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a DELETE route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -379,21 +492,32 @@ def delete(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.DELETE, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.DELETE,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def options(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
-) -> Path:
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add an OPTIONS route.
 
     Args:
         path_or_route: The path to this route, or the route itself.
         doc: The description of the route to be used in documentation.
         cache_rate: Reload the cache for this route every x number of requests. `-1` means to never cache.
+        steps: Build steps to run before this route is executed.
+        parallel_build: Whether to run the build steps in parallel. If this is not specified, this is defined by the app configuration.
 
     Example:
         ```py
@@ -404,15 +528,33 @@ def options(
             return "Hello, view.py!"
         ```
     """
-    return _method_wrapper(path_or_route, doc, Method.OPTIONS, cache_rate)
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.OPTIONS,
+        cache_rate,
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 def websocket(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
-) -> Path:
+    *,
+    steps: Iterable[str] | None = None,
+    parallel_build: bool | None = _DefinedByConfig,
+) -> Path[P]:
     """Add a websocket route."""
-    return _method_wrapper(path_or_route, doc, Method.WEBSOCKET, -1, inputs=[2])
+    return _method_wrapper(
+        path_or_route,
+        doc,
+        Method.WEBSOCKET,
+        -1,
+        inputs=[2],
+        steps=steps,
+        parallel_build=parallel_build,
+    )
 
 
 _STR_METHOD_MAPPING: dict[StrMethod, Method] = {
@@ -427,12 +569,13 @@ _STR_METHOD_MAPPING: dict[StrMethod, Method] = {
 
 
 def route(
-    path_or_route: str | None | RouteOrCallable = None,
+    path_or_route: str | None | RouteOrCallable[P] = None,
     doc: str | None = None,
     *,
     cache_rate: int = -1,
     methods: Iterable[StrMethod] | None = None,
-) -> Path:
+    steps: Iterable[str] | None = None,
+) -> Path[P]:
     """Add a route that can be called with any method (or only specific methods).
 
     Args:
@@ -455,7 +598,10 @@ def route(
         doc,
         None,
         cache_rate,
-        method_list=[_STR_METHOD_MAPPING[i] for i in methods] if methods else None,
+        method_list=[_STR_METHOD_MAPPING[i] for i in methods]
+        if methods
+        else None,
+        steps=steps,
     )