+++ tests/test_build.py
@@ -1,0 +1,8 @@
+from ward import test
+
+from view import new_app
+
+
+@test("test static exports")
+def test_static_exports():
+    ...